Add lib2to3 from Python 3.8 RC1
All references have been updated to root in crosswind so that we don't
accidentally include the system level lib2to3. Some tests had to have
some asserts commented out due to their reliance on capturing stderr which
pytest already does. They will be updated to use pytest's stderr capture
in the near future (but not as this body of work).