Get CI going with linting and testing
This requires a non-trivial amount of changes including removing
old files, fixing several linting errors (the CI pipeline will only
worry about actual errors for now) which includes a few suboptimal
fixes, and so on.

This also updates the README with information about using the Makefile
and how to run the current dev version of crosswind.