Rearrange fixers into separate package
The intent here is to move towards being able to ship custom
"fixer suits" in a similar manner to how pytest can do its plugins.
Ideally we can automate discovery in the future.

This is a large changeset but is almost entirely just the moving
of files and updating imports and paths to reflect the new organization.